+++ src/main/java/okeanos/restApp/resources/SaisonResource.java
@@ -18,32 +18,42 @@ public class SaisonResource extends AbstractResource {
 
 		get(ressourcePath, (request, response) -> {
 			setSecurity(request, response);
+			if (!SecurityResource.isLogin(request)) {
+				throw new IllegalAccessException("Illegal Access");
+			}
 			return JsonUtil.toJson(SaisonDao.getAllItems());
 		});
 
 		get(ressourcePath + "/currentSaison", (request, response) -> {
 			setSecurity(request, response);
+			if (!SecurityResource.isLogin(request)) {
+				throw new IllegalAccessException("Illegal Access");
+			}
 			return JsonUtil.toJson(SaisonDao.getCurrentSaison());
 		});
 
 		get(ressourcePath + "/:id", (request, response) -> {
 			setSecurity(request, response);
+			if (!SecurityResource.isLogin(request)) {
+				throw new IllegalAccessException("Illegal Access");
+			}
 			return JsonUtil.toJson(SaisonDao.getItemById(new Long(request.params(":id"))));
 		});
 
 		delete(ressourcePath + "/:id", (request, response) -> {
 			setSecurity(request, response);
+			if (!SecurityResource.isAdmin(request)) {
+				throw new IllegalAccessException("Illegal Access");
+			}
 			return JsonUtil.toJson(SaisonDao.deleteItem(new Long(request.params(":id"))));
 		});
 
 		post(ressourcePath, (request, response) -> {
 			setSecurity(request, response);
-			Saison saison = gson.fromJson(request.body(), Saison.class);
 			if (!SecurityResource.isAdmin(request)) {
-				// Modification non authorisée, on retourne l'objet initial
-				return JsonUtil.toJson(SaisonDao.getItemById(saison.getId()));
+				throw new IllegalAccessException("Illegal Access");
 			}
-			return JsonUtil.toJson(SaisonDao.save(saison));
+			return JsonUtil.toJson(SaisonDao.save(gson.fromJson(request.body(), Saison.class)));
 		});
 	}
